server/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
psycopg2
databases
psycopg2-binary
alembic
python-dotenv
argon2-cffi
cryptography
python-jose[cryptography]
tenacity
python-multipart
python-socketio
websockets
asyncio
